+++ R/cosmos_data.R
@@ -7,9 +7,8 @@
 #' @param meta_network Prior knowledge network (PKN).  By default COSMOS use a 
 #'   PKN derived from Omnipath, STITCHdb and Recon3D. See details on the data 
 #'   \code{\link{meta_network}}.
-#' @param tf_regulon Collection of transcription factor - target interactions.
-#'   A default collection from dorothea can be obtained by the 
-#'   \code{\link{load_tf_regulon_dorothea}} function.
+#' @param tf_regulon Collection of transcription factor-target interactions with
+#'   columns `tf`, `sign`, and `target`.
 #' @param signaling_data Numerical vector, where names are signaling nodes 
 #'   in the PKN and values are from \{1, 0, -1\}.  Continuous data will be 
 #'   discretized using the \code{\link{sign}} function.  
@@ -103,9 +102,8 @@ validate_cosmos_data <- function(x){
 #' @param meta_network Prior knowledge network (PKN).  By default COSMOS use a
 #'   PKN derived from Omnipath, STITCHdb and Recon3D.  See details on the data 
 #'   \code{\link{meta_network}}.
-#' @param tf_regulon Collection of transcription factor - target interactions.  
-#'   A default collection from dorothea can be obtained by the 
-#'   \code{\link{load_tf_regulon_dorothea}} function.
+#' @param tf_regulon Optional collection of transcription factor-target
+#'   interactions with columns `tf`, `sign`, and `target`.
 #' @param signaling_data Numerical vector, where names are signaling nodes 
 #'   in the PKN and values are from \{1, 0, -1\}.  Continuous data will be 
 #'   discretized using the \code{\link{sign}} function.