+++ R/run_COSMOS_signaling_to_metabolism.R
@@ -31,7 +31,9 @@
 #' data(toy_signaling_input)
 #' data(toy_metabolic_input)
 #' data(toy_RNA)
+#' tf_regulon <- data.frame(tf = "MYC", sign = 1, target = "SLC2A1")
 #' test_for <- preprocess_COSMOS_signaling_to_metabolism(meta_network = toy_network,
+#'                              tf_regulon = tf_regulon,
 #'                              signaling_data = toy_signaling_input,
 #'                              metabolic_data = toy_metabolic_input,
 #'                              diff_expression_data = toy_RNA,
@@ -69,4 +71,3 @@ validate_cosmos_data_signaling_to_metabolism <- function(data){
         stop("missing inputs detected. Input data should be obtained by running preprocess_cosmos_signaling_to_metabolism.")
     
 }
-