+++ tests/testthat/test-run_COSMOS_signaling_to_metabolism.R
@@ -12,6 +12,7 @@ test_that("test run COSMOS signaling to metabolism", {
   
   res <- preprocess_COSMOS_signaling_to_metabolism(signaling_data = signaling_data,
                            meta_network = meta_network,
+                           tf_regulon = test_tf_regulon(),
                            metabolic_data = metabolic_data,
                            diff_expression_data = expression_data,
                            remove_unexpressed_nodes = FALSE,